pyproject.toml
[project]
name = "icair-get-dp"
version = "0.1.0"
description = "ICAIR 2025 Conference Profile Picture Generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.0.0",
    "gunicorn>=21.0.0",
    "pillow-heif>=0.13.0"
]
